fix: move shared code to separate module
