+++ apps/server/scripts/benchmarks/serialize-state-update/plot.ts
@@ -1,35 +1,32 @@
 import { Logger } from "@mp/logger";
-
 import yargs from "yargs";
 import { hideBin } from "yargs/helpers";
 import {
-  average,
   benchmark,
-  durationString,
-  implementationNames,
   sizeString,
+  durationString,
+  average,
   total,
-} from "./once";
+  implementationNames,
+} from "./shared";
 
-if (require.main === module) {
-  const logger = new Logger(console);
-  const { begin, end, offset, iterations, implementation } = readCliOptions();
+const logger = new Logger(console);
+const { begin, end, offset, iterations, implementation } = readCliOptions();
 
+logger.info(
+  `Benchmarking ${implementation} at ${iterations} iterations from ${begin} to ${end} with offset ${offset}`,
+);
+
+for (let n = begin; n <= end; n += offset) {
+  const { times, size } = benchmark(n, iterations, implementation);
   logger.info(
-    `Benchmarking ${implementation} at ${iterations} iterations from ${begin} to ${end} with offset ${offset}`,
+    [
+      `N: ${n}`,
+      `state: ${sizeString(size)}`,
+      `time (avg): ${durationString(average(times))}`,
+      `time (total): ${durationString(total(times))}`,
+    ].join("\t"),
   );
-
-  for (let n = begin; n <= end; n += offset) {
-    const { times, size } = benchmark(n, iterations, implementation);
-    logger.info(
-      [
-        `N: ${n}`,
-        `state: ${sizeString(size)}`,
-        `time (avg): ${durationString(average(times))}`,
-        `time (total): ${durationString(total(times))}`,
-      ].join("\t"),
-    );
-  }
 }
 
 function readCliOptions(argv = process.argv) {
